+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Type hints for the complete public interface - [#36](https://github.com/hivesolutions/netius/issues/36)
 * Support for the SRV, SVCB, HTTPS and CAA DNS record types
 * Support for serving ASGI applications, such as the ones written with Starlette or FastAPI - [#85](https://github.com/hivesolutions/netius/issues/85)
+* Support for the `proxy.regex` Consul tag, allowing regex based routing rules to be registered per service - [#109](https://github.com/hivesolutions/netius/issues/109)
 
 ### Changed